Revisions

+++ scripts/get_mysql_data.sh
@@ -8,10 +8,28 @@ cd data
 #
 echo "MySQL Example Databases"
 echo ".. world - InnoDB"
-curl --silent -o world-innodb.sql.gz http://downloads.mysql.com/docs/world_innodb.sql.gz
+[ ! -f world-innodb.sql.gz ] && curl --silent -o world-innodb.sql.gz http://downloads.mysql.com/docs/world_innodb.sql.gz
 echo ".. world - MyISAM"
-curl --silent -o world-myisam.gz http://downloads.mysql.com/docs/world.sql.gz
+[ ! -f world-myisam.sql.gz ] && curl --silent -o world-myisam.sql.gz http://downloads.mysql.com/docs/world.sql.gz
 echo ".. world - Sakila"
-curl --silent -o sakila-db.sql.gz http://downloads.mysql.com/docs/sakila-db.tar.gz
+[ ! -f sakila-db.tar.gz ] && curl --silent -o sakila-db.tar.gz http://downloads.mysql.com/docs/sakila-db.tar.gz
 echo ".. world - Employee"
-curl --silent -o employee.tar.bz2 https://launchpadlibrarian.net/24493789/employees_db-dump-files-1.0.5.tar.bz2
+[ ! -f employee.tar.bz2 ] && curl --silent -o employee.tar.bz2 https://launchpadlibrarian.net/24493789/employees_db-dump-files-1.0.5.tar.bz2
+[ ! -f employee-code.tar.bz2 ] && curl -o employees-code.tar.bz2 https://launchpadlibrarian.net/24493350/employees_db-code-1.0.6.tar.bz2
+
+tar xvfz *.tar.gz
+gunzip *.gz
+tar xvfj *.bz2
+
+mysqladmin create world_myisam
+mysql world_myisam < world-myisam.sql
+mysqladmin create world_innodb
+mysql world_innodb < world-innodb.sql
+cd sakila-db
+mysql < sakila-schema.sql
+mysql < sakila-data.sql
+
+cd ..
+cd employees_db
+mysql < employees.sql
+exit 0
